first step into a strong architecture
